fix(WR): fix wr capture for WR/PGTK
